update to polyfill
